Clean up compiler warnings: dead variable, missing narrowing cast
- scenes_advanced.h: remove an unused double px,py,pz,dx,dy,dz,ex,ey,ez
  declaration in BilinearPatch's hit() - dead since the actual dpdu
  computation uses the float-typed blp_point() output instead.
- sampling_patched.h: NormalPDF() returns float but called the
  double-returning Gaussian() without a cast (C4244 truncation warning);
  add the explicit cast, matching the pattern already used in
  scalar_distributions.h's own NormalPDF.

+++ src/shared/sampling_patched.h
@@ -1032,7 +1032,7 @@ CPU_GPU float InvertTrimmedExponentialSample(float x, float c, float xMax) {
 // Normal (Gaussian) distribution
 // pbrt-v4: NormalPDF / SampleNormal / InvertNormalSample / SampleTwoNormal
 CPU_GPU float NormalPDF(float x, float mu = 0.f, float sigma = 1.f) {
-	return Gaussian(x, mu, sigma);
+	return static_cast<float>(Gaussian(x, mu, sigma));
 }
 
 CPU_GPU float SampleNormal(float u, float mu = 0.f, float sigma = 1.f) {
